I took the liberty of adding the code for pre-login script.  As well as modifying BINDIR from /bin to /.bin to hide the drivers.
Also added script lines to create a shortcut in /var/lib/gdm which is the home directory of user gdm, pointing to the driver in the users BINDIR.  So only one copy is needed of the file.  (Otherwise 2 copies would be needed, and since we are on a tablet with small disc space I felt the extra code was necessary).  I've tested most aspects of my added code, but have not actually tried it on a fresh system.  Let me know what you think.